TM1637 upto L6, and reordering of classes in calls/includes etc

+++ main/L4_DomainModels/I2c/IoPins/IoPins.cpp
@@ -28,7 +28,7 @@ Joystick& IoPins::GetJoystick(Types::EJoystickId joystickId)
 
 Led& IoPins::GetLed(Types::ELedId ledId)
 {
-    return _leds[ledId];
+    return _leds[static_cast<uint8_t>(ledId)];
 }
 
 SystemButton& IoPins::GetSystemButton()
